+++ lib/src/bloc/src/use_case_builders/use_case_builder.dart
@@ -1,4 +1,5 @@
 export 'src/event_subscription.dart';
 export 'src/relay_use_case_builder.dart';
+export 'src/state_relay.dart';
 export 'src/stateful_use_case_builder.dart';
 export 'src/use_case_builder.dart';